Improve developer experience for web platform tests
* Fix process shutdown on Windows to work consistently.

* Filter the logs a bit to make them more readable.

* Don't require manually marking .worker/.sharedworker/serviceworker variants as skipped in to-run.yaml.
