+++ test/web-platform-tests/run-tuwpts.js
@@ -3,7 +3,7 @@ const path = require("path");
 const { describe, before, after } = require("mocha-sugar-free");
 const { spawnSync } = require("child_process");
 const { readManifest, getPossibleTestFilePaths } = require("./wpt-manifest-utils.js");
-const startWPTServer = require("./start-wpt-server.js");
+const wptServer = require("./wpt-server.js");
 
 const wptPath = path.resolve(__dirname, "tests");
 const testsPath = path.resolve(__dirname, "to-upstream");
@@ -20,14 +20,14 @@ const possibleTestFilePaths = getPossibleTestFilePaths(manifest);
 
 let wptServerURL, serverProcess;
 const runSingleWPT = require("./run-single-wpt.js")(() => wptServerURL);
-before({ timeout: 30 * 1000 }, async () => {
-  const { urls, subprocess } = await startWPTServer({ toUpstream: true });
+before({ timeout: 30_000 }, async () => {
+  const { urls, subprocess } = await wptServer.start({ toUpstream: true });
   wptServerURL = urls[0];
   serverProcess = subprocess;
 });
 
 after(() => {
-  serverProcess.kill();
+  wptServer.kill(serverProcess);
 });
 
 describe("Local tests in web-platform-test format (to-upstream)", () => {
