+++ test/web-platform-tests/to-run.yaml
@@ -1402,11 +1402,9 @@ idlharness.any.html: [fail, Depends on fetch]
 percent-encoding.window.html: [fail, Depends on fetch]
 toascii.window.html: [fail, Depends on fetch]
 url-constructor.any.html**: [fail, Depends on fetch]
-url-constructor.any.worker.html*: [fail, Depends on Worker]
 url-origin.any.html: [fail, Depends on fetch]
 url-setters-a-area.window.html**: [fail, Depends on fetch]
 url-setters.any.html**: [fail, Depends on fetch]
-url-setters.any.worker.html*: [fail, Depends on Worker]
 urlencoded-parser.any.html: [fail, Depends on fetch]
 
 ---
@@ -1414,13 +1412,9 @@ urlencoded-parser.any.html: [fail, Depends on fetch]
 DIR: websockets
 
 "**/**?wpt_flags=h2": [fail-slow, HTTP/2 web sockets not implemented]
-"*.any.sharedworker.html?wss": [fail-slow, Needs SharedWorker implementation]
-"*.any.worker.html?wss": [fail-slow, Needs Worker implementation]
 Create-blocked-port.any.html: [fail-slow, Not implemented]
 Create-blocked-port.any.html?wss: [fail-slow, Not implemented]
 Create-on-worker-shutdown.any.html: [fail, Needs Worker implementation]
-Send-data.worker.html?wss: [fail-slow, Needs Worker implementation]
-basic-auth.any.serviceworker.html?wss: [fail-slow, Unknown]
 cookies/third-party-cookie-accepted.https.html: [fail, 'https://github.com/salesforce/tough-cookie/issues/80']
 idlharness.any.html: [fail, Depends on fetch]
 interfaces/WebSocket/close/close-connecting.html*: [fail, Potentially buggy test as Chrome fails it too]
